requirements.txt
asgiref==3.6.0
attrs==22.2.0
build==0.10.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
packaging==23.0
Pillow==10.0.0
pluggy==1.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pyproject_hooks==1.0.0
pytest==7.2.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
